GH-10 Refactor: move model calc to different class
This just moves the code for calculating the parameters (or getting them from the DB) to the catmodel_info class.
